+++ src/login.ts
@@ -1,35 +1,48 @@
+import { AuthenticationDetails } from 'amazon-cognito-identity-js';
+import { Unauthorised } from 'aws-lambda-response-helper';
+import { IAuthResponse, ILoginRequest } from './models';
+import { getCognitoUser } from './internal';
 
-export async function attemptLogin(request: ILoginRequest, authDetails: AuthenticationDetails, userType: UserTypes): Promise<IAuthResponse> {
-    const cognitoUser = getCognitoUser(request.username, userType);
+export async function attemptLogin(
+    loginRequest: ILoginRequest,
+    clientId: string,
+    poolId: string): Promise<IAuthResponse> {
+    const authDetails = new AuthenticationDetails({
+        Username: loginRequest.username,
+        Password: loginRequest.password
+    });
+    const cognitoUser = getCognitoUser(loginRequest.username, clientId, poolId);
 
     return new Promise((resolve, reject) => {
         cognitoUser.authenticateUser(authDetails, {
             onSuccess: (data) => {
                 resolve({
-                    user_type: userType,
-                    api_url: userType === UserTypes.ADMIN ? process.env.ADMIN_API_URL! : process.env.CLIENT_API_URL!,
-                    token: data.getIdToken().getJwtToken(),
-                    refresh_token: data.getRefreshToken().getToken()
+                    idToken: data.getIdToken().getJwtToken(),
+                    refreshToken: data.getRefreshToken().getToken()
                 });
             },
             onFailure: (err) => {
-                reject(err);
+                if (err.code && (err.code === 'UserNotFoundException' || err.code === 'NotAuthorizedException')) {
+                    reject(new Unauthorised());
+                } else {
+                    reject(err);
+                }
             },
             newPasswordRequired: (userAttributes, requiredAttributes) => {
-                if (request.new_password) {
+                if (loginRequest.new_password) {
                     delete userAttributes.email_verified;
                     requiredAttributes = userAttributes;
-                    cognitoUser.completeNewPasswordChallenge(request.new_password, requiredAttributes, {
+
+                    cognitoUser.completeNewPasswordChallenge(loginRequest.new_password, requiredAttributes, {
                         onSuccess: async (_data) => {
-                            resolve(await attemptLogin(request, authDetails, userType));
+                            resolve(await attemptLogin(loginRequest, clientId, poolId));
                         },
                         onFailure: (err) => {
                             reject(err);
                         }
                     });
                 } else {
-                    response.new_password_required = true;
-                    resolve(response);
+                    reject(new Error('New password required'));
                 }
             }
         });